Move Blueprints v1-specific files to v1 subdirectory
