+++ packages/playground/blueprints/src/lib/steps/install-plugin.ts
@@ -4,7 +4,7 @@ import { installAsset } from './install-asset';
 import { activatePlugin } from './activate-plugin';
 import { writeFile } from './write-file';
 import { zipNameToHumanName } from '../utils/zip-name-to-human-name';
-import type { Directory } from '../resources';
+import type { Directory } from '../v1/resources';
 import { joinPaths } from '@php-wasm/util';
 import { writeFiles } from '@php-wasm/universal';
 import { logger } from '@php-wasm/logger';
