+++ packages/playground/blueprints/src/lib/v1/resources.ts
@@ -12,10 +12,10 @@ import {
 	resolveCommitHash,
 	sparseCheckout,
 } from '@wp-playground/storage';
-import { zipNameToHumanName } from './utils/zip-name-to-human-name';
+import { zipNameToHumanName } from '../utils/zip-name-to-human-name';
 import { fetchWithCorsProxy } from '@php-wasm/web';
 import { StreamedFile } from '@php-wasm/stream-compression';
-import type { StreamBundledFile } from './blueprint';
+import type { StreamBundledFile } from './types';
 
 export type { FileTree };
 export const ResourceTypes = [
