+++ packages/playground/blueprints/src/lib/v1/types.ts
@@ -1,5 +1,5 @@
 import type { SupportedPHPVersion } from '@php-wasm/universal';
-import type { StepDefinition } from './steps';
+import type { StepDefinition } from '../steps';
 import type { FileReference } from './resources';
 import type { StreamedFile } from '@php-wasm/stream-compression';
 import type { Filesystem } from '@wp-playground/storage';
